refactor(#21) :: 지도 스크린 컴포넌트화
