+++ src/components/Maps/ScreenMap/style.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const ScreenMap_Layout = styled.div`
+  width: 100%;
+  height: 100%;
+`;
